refactor: rename session package to auth
