+++ internal/auth/login.go
@@ -1,4 +1,4 @@
-package session
+package auth
 
 import (
 	"context"
